finishing project

+++ sql/create_payloads_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS public.payloads_raw_data
     id TEXT ,
     name TEXT,    
     launch TEXT,
-    mass_kg NUMERIC,
-    mass_lbs NUMERIC,
+    mass_kg BIGINT,
+    mass_lbs BIGINT,
     inserted_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
